added filters to broadcast; increased number of test data for broadcast;
