+++ src/main/java/global/govstack/threat_service/controller/impl/BroadcastControllerImpl.java
@@ -2,9 +2,8 @@ package global.govstack.threat_service.controller.impl;
 
 import global.govstack.threat_service.controller.BroadcastControllerInterface;
 import global.govstack.threat_service.controller.exception.NotFoundException;
-import global.govstack.threat_service.dto.BroadcastCreateDto;
-import global.govstack.threat_service.dto.BroadcastDto;
-import global.govstack.threat_service.repository.entity.ThreatEvent;
+import global.govstack.threat_service.dto.broadcast.BroadcastCreateDto;
+import global.govstack.threat_service.dto.broadcast.BroadcastDto;
 import global.govstack.threat_service.service.BroadcastService;
 import global.govstack.threat_service.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -26,8 +25,8 @@ public class BroadcastControllerImpl implements BroadcastControllerInterface {
     private final UserService userService;
 
     @Override
-    public Page<BroadcastDto> getAllBroadcasts(Pageable pageable) {
-        return broadcastService.getAllBroadcasts(pageable);
+    public Page<BroadcastDto> getAllBroadcasts(String country, boolean active, UUID userUUID, Pageable pageable) {
+        return broadcastService.getAllBroadcasts(country, active, userUUID, pageable);
     }
 
     @Override
