+++ src/main/java/global/govstack/threat_service/dto/DemoDto.java
@@ -1,5 +1,7 @@
 package global.govstack.threat_service.dto;
 
+import global.govstack.threat_service.dto.threat.IncomingThreatDto;
+
 public record DemoDto(
     String key,
     IncomingThreatDto threat
