+++ src/main/java/global/govstack/threat_service/dto/broadcast/CreateBroadcastCountyDto.java
@@ -1,4 +1,4 @@
-package global.govstack.threat_service.dto;
+package global.govstack.threat_service.dto.broadcast;
 
 
 public record CreateBroadcastCountyDto(
