+++ src/main/java/global/govstack/threat_service/dto/broadcast/KafkaBroadcastDto.java
@@ -1,4 +1,4 @@
-package global.govstack.threat_service.dto;
+package global.govstack.threat_service.dto.broadcast;
 
 import lombok.Builder;
