+++ src/main/java/global/govstack/threat_service/mapper/BroadcastMapper.java
@@ -1,7 +1,7 @@
 package global.govstack.threat_service.mapper;
 
-import global.govstack.threat_service.dto.BroadcastCreateDto;
-import global.govstack.threat_service.dto.BroadcastDto;
+import global.govstack.threat_service.dto.broadcast.BroadcastCreateDto;
+import global.govstack.threat_service.dto.broadcast.BroadcastDto;
 import global.govstack.threat_service.repository.entity.Broadcast;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
