+++ src/main/java/global/govstack/threat_service/service/ThreatService.java
@@ -3,8 +3,8 @@ package global.govstack.threat_service.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import global.govstack.threat_service.controller.exception.InternalServerException;
-import global.govstack.threat_service.dto.KafkaThreatDto;
-import global.govstack.threat_service.dto.ThreatDto;
+import global.govstack.threat_service.dto.threat.KafkaThreatDto;
+import global.govstack.threat_service.dto.threat.ThreatDto;
 import global.govstack.threat_service.dto.overview.OverviewThreatDto;
 import global.govstack.threat_service.mapper.ThreatMapper;
 import global.govstack.threat_service.repository.CountryThreatRepository;
